Working with new window methods
