+++ src/client/public/config/openfin/local.app.json
@@ -1,35 +1,35 @@
 {
-  "devtools_port": 9090,
-  "splashScreenImage": "http://localhost:3000/static/media/splash-screen.jpg",
-  "startup_app": {
-    "name": "Reactive Trader Cloud (LOCAL)",
-    "url": "http://localhost:3000/",
-    "uuid": "reactive-trader-cloud-local",
-    "applicationIcon": "http://localhost:3000/static/media/adaptive-mark-large.png",
-    "autoShow": true,
-    "defaultWidth": 1280,
-    "defaultHeight": 900,
-    "minWidth": 800,
-    "minHeight": 600,
-    "resizable": true,
-    "maximizable": true,
-    "frame": false
-  },
-  "runtime": {
-    "arguments": "--force-device-scale-factor=1",
-    "version": "9.61.33.32",
-    "futureVersion": "9.61.33.32",
-    "forceLatest": true
-  },
-  "shortcut": {
-    "company": "Adaptive Consulting",
-    "description": "Desktop Reactive Trader Cloud",
-    "icon": "http://localhost:3000/static/media/icon.ico",
-    "name": "Reactive Trader Cloud (LOCAL)"
-  },
-  "services": [
-    {
-      "name": "layouts"
-    }
-  ]
+    "devtools_port": 9090,
+    "splashScreenImage": "http://localhost:3000/static/media/splash-screen.jpg",
+    "startup_app": {
+        "name": "Reactive Trader Cloud (LOCAL)",
+        "url": "http://localhost:3000/",
+        "uuid": "reactive-trader-cloud-local",
+        "applicationIcon": "http://localhost:3000/static/media/adaptive-mark-large.png",
+        "autoShow": true,
+        "defaultWidth": 1280,
+        "defaultHeight": 900,
+        "minWidth": 800,
+        "minHeight": 600,
+        "resizable": true,
+        "maximizable": true,
+        "frame": false
+    },
+    "runtime": {
+        "arguments": "--force-device-scale-factor=1",
+        "version": "9.61.33.32",
+        "futureVersion": "9.61.33.32",
+        "forceLatest": true
+    },
+    "shortcut": {
+        "company": "Adaptive Consulting",
+        "description": "Desktop Reactive Trader Cloud",
+        "icon": "http://localhost:3000/static/media/icon.ico",
+        "name": "Reactive Trader Cloud (LOCAL)"
+    },
+    "services": [
+        {
+            "name": "layouts"
+        }
+    ]
 }
